backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.28.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
